mcp_tools/citation_checker/requirements.txt
fastmcp>=2.0
httpx>=0.27
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.24
respx>=0.21
